pyproject.toml
[project]
name = "bestmlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kristof Drexler", email = "kristof.drexler@kauza.dk" }
]
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.35.0",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "transformers>=4.53.0",
]

[project.scripts]
bestmlops = "bestmlops:main"
local_deploy = "bestmlops.local_deploy:deploy_model"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "ruff>=0.12.1",
]
